uk/plat: Add flags parameter to ukplat_lcpu_start
We have a flags parameter for future extensibility for the
run function. This commit adds a flags parameter also to
the start function to similarly provide the possibility for
future extensions, for example, to filter CPUs with a certain
characteristic (big.LITTLE) or put the target CPU into an
architecture-specific startup state.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Michalis Pappas <michalis.pappas@opensynergy.com>
Reviewed-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Approved-by: Michalis Pappas <michalis.pappas@opensynergy.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #469