uk/plat: Make num parameter in/out
In the current API, we don't have a way to retry operations if they
failed because we don't know for which CPU the operation failed. This
commit changes the num parameter so that it can function as a
in/out parameter and receives the number of CPUs on which the
operation was carried out successfully. The API then stop at the
first CPU that fails.

The commit also adds const qualifiers where appropriate.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>