uk/plat: Remove list entry from ukplat_lcpu_func
Using a linked list to queue function objects requires that we create
copies of the function objects so that we can link them to the
(multiple) target CPU function lists. That also means, we have to do
dynamic memory allocation. As this might not be appropriate in
every case, we prefer to draw from a static pool that does not
require linking.

We also make the regs function parameter optional for now as not all
platforms might support providing a trap frame.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>