plat/drivers/gic-v2: Add asserts
Some of the parameters allow larger value ranges than accepted by the
GIC. We therefore add additional assertations to make sure values are
valid. We also replace the crash in `gic_set_irq_target` as specifying a
wrong IRQ number can be considered a bug instead of a runtime error.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>