lib/ukstore: Add dynamic entry registration
This commit adds the possibility to register entries and folders at
runtime. Each library will register folders to itself. Entries can
be registered to the folders afterwards if the name is known.

The entries contain the getters/setters used in passing information.
Overall, the dynamic API provides calls to register, free and call
entries.

The dynamic version calls of the library should be enclosed in
ifdefs to ensure that everything works correctly.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>