plat/kvm/x86: Refactor CPU boot trampoline
The current boot path is based on the assumption that the processor
is starting in a multiboot environment. However, this does not need
to be the case, for example, when running with Firecracker which uses
the Linux boot protocol. Furthermore, when booting secondary cores in
a multi-processor setup, these cores start in 16-bit real mode and need
a different initialization path.

Obviously, we do not want to introduce and maintain multiple ASM
boot trampolines for the various combinations of VMMs and SP/MP
setups. This commit therefore separates the actual initialization of the
processor from the boot environment (e.g., multiboot) and adds further
entry points depending on the architectural state (e.g., 16-bit mode).

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Signed-off-by: cristian-vijelie <cristianvijelie@gmail.com>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #502