plat/[kvm|xen]/x86: Refactor trap initialization
This commit refactors how traps are setup on x86 KVM. This is necessary
to support MP operation. We are currently using dedicated stacks for
IRQs, traps, and some critical traps (e.g., double fault). However, we
only have one global instance for each of these stacks which does
not work with multiprocessor configurations, where each processor
needs its own stacks. We thus extend the code to offer stacks for each
processor.

However, we do not keep all of the previous stacks but only use one
stack for critical traps and one stack for IRQs. All other traps operate
on whatever stack is currently configured. This saves memory and
allows nested traps (e.g., when triggering a page fault for memory
probing while handling another trap -> used for debugging).
Using IST in the same way is usually also what other OSes (e.g., Windows
and Linux) do. A problem with this approach is that it causes a double
fault whenever there is not enough space left on the stack to push the
stack frame. General-purpose OSes do not have this problem, because
they can have an implicit stack switch during mode switch from user
to kernel mode. But in case we run into a double fault, we will switch
to the dedicated stack and in the future can try to recover from this
error (e.g., by extending the current stack). Until then, we can
assume that as a Unikernel we know our workload better than other
OSs and are thus able to scale the stacks more appropriately for the
workload at hand.

Besides incorporating some general cleanup, this commit also adds
proper switching of all segment registers when loading a new GDT.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #502