plat/common: Add temporary APIC IRQ ack
On x86 we currently use a hybrid setup with PIC and APIC both
active. As IRQs need to be acknowleged differently depending
on if they have been raised by the PIC or the APIC, we just ack
SMP-related IRQs in the APIC here temporarily until we have
completely moved to the APIC also for I/O devices.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>