lib/posix-user: Harmonize group and user code
The user/passwd management functions currently use the
iterator interface also internally. However, as we only have
a single user, this is a waste of cycles. We thus do it like
for groups and make the default passwd entry static global
so we can work directly with it. The commit also changes
variable names to make passwd and group code more
similar.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #504