lib/posix-user: Fix getgrnam_r and add getgrgid_r
The current implementation of getgrnam_r has a bug as it does not
reserve memory for the actual array but only for the group names.
This commit fixes the bug and generalizes the implementation by
moving it do a dedicated function. This way, we can easily add
getgrgid_r.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #504