plat/common: Add option to enforce W^X
Memory protections are set differently for the various platforms
depending on how the static page table is constructed. This commit
adds the option to enforce correct memory protections using the
paging API so that writeable memory is not executable and vice
versa. This is important to prevent code injection attacks like using
a simple buffer overflow on the stack to inject shellcode or use
a write primitive to modify the .text segment.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Signed-off-by: Michalis Pappas <michalis.pappas@opensynergy.com>
Reviewed-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #506