lib/vfscore: Add getdents64 system call
Currently `getdents` syscall is not properly implemented as it assumes
the layout `struct dirent` structure is the same as the one of `struct
dirent64`. The structures are different, so an implementation of
`getdents64` is required.

This commit adds the implementation, required macros, and build
information for the `getdents64` syscall. Implementation is a copy of
the `getdents` syscall.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>