ofw: Fix fdt_get_address, get cell info from parent node.
As stated in DT Spec v0.2, section 2.3.5 "#address-cells and #size-cells",
the properties #address-cells and #size-cells need to be read from the
parent node, not the node itself.
Every parent node is required to have this information in a compliant DT.

This effectively reverts the function back to its introduction in 42c1127.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>