lib/syscall_shim: Ignore only unnumbered legacy syscalls
Currently, all the unnumbered system calls are ignored
during the build process. This shouldn't be the case
for architecture specific syscalls. Those should generate
the failure of the build process if their number is not defined
for the corresponding architecture (see arch_prctl).
This commit introduces a separate header file that allows
us to only ignore unnumbered legacy syscalls.

Signed-off-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #509