lib/posix-user: Add implementation of s/getgroups
This commit changes the stubbed getgroups/setgroups to do more
meaningful operations. Since we only have a single group and this group
does not need to be returned in the list, we can effectively
ignore supplementary groups. However, we ensure that when setting
supplementary groups, only the default GID (i.e., a valid GID) is
attempted to be set so it is fine for getgroups to just return 0.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>