build: Generate uk_clean_list in bulk.
Writes all files to clean of a library into the uk_clean_list in bulk
and use sed to turn spaces into linebreaks, while trimming leading and
tailing spaces.

This produces the same uk_clean_list as before, but can be lot faster
when building libraries with a lot of sources.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>