plat/kvm/arm: Do not limit the vaddr size when paging is enabled
While setting up the MMU, the platform sets vaddr size to the be equal
to paddr size. This is done in anticipation that the whole address space
is mapped 1-to-1. When paging is enabled this assumption is no longer
valid, as we use a 48-bit virtual address space regardless of the paddr
width. In that case just use the default values defined for paging.

Signed-off-by: Michalis Pappas <michalis.pappas@opensynergy.com>