lib/syscall_shim: TLS pointer of caller
If `CONFIG_LIBSYSCALL_SHIM_HANDLER_ULTLS` is configured, system call
handlers can retrieve the caller's TLS pointer with
`uk_syscall_ultlsp()` if the system call was called by a binary system
call.
Please note that only if the system call was called by a binary system
call and the TLS pointer of the caller is different to the Unikraft-TLS
pointer, this function will return the caller's TLS pointer.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>