lib/posix-process: clone: Support userland TLSes
REWRITE: `_clone()` assumes that a passed TLS pointer is an Unikraft TLS.
The only exception exists if `_clone()` is called from a context
where a custom TLS is already active (depends on
`CONFIG_LIBSYSCALL_SHIM_HANDLER_ULTLS`). In such a case, an
Unikraft TLS is allocated but the passed TLS pointer is activated.
The reason is that Unikraft libraries place TLS variables and use
the TLS effectively as TCB.
In case no TLS is handed over (CLONE_SETTLS is not set), _clone will
still allocate an Unikraft TLS but sets the TLS architecture pointer
to zero.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>