arch/arm64: Add processor-generated randomness (FEAT_RNG)
FEAT_RNG is introduced as an OPTIONAL feature in Armv8.5-a, and specifies
a hardware-based RNG that is implemented as a cryptographically secure
Deterministic Random Bit Generator (DRBG), seeded by a TRNG.

The RNG can be accessed by the RNDR and RNDRSS registers. RNDR reseeds the
DRBG at an IMPLEMENTATION DEFINED rate. RNDRRS register that seeds the
DRBG before generating the next random number.

This commit adds a new Kconfig option to enable FEAT_RNG and implements
the ukarch_random API using RNDR / RNDRRS.

Notice that as the entropy generated by a TRNG is limited, the caller
must always check the returned value to ensure that a random number was
successfully generated.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #434