plat/virtio_pci: Fix buffer length argument in function call
In the `vpci_legacy_pci_config_get` function, `len` is equal to the
number of elements in the buffer, where each element is of size
`type_len` bytes.

In this routine, the `_virtio_cread_bytes` is called using these
two parameters if an entity has less then 4 bytes. However, when calling
`virtio_cread_bytes_many`, the function has only one argument for
specifying the buffer length, and it should be used to pass the total
size of the buffer in bytes (i.e. len * type_len), not the
number of elements.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>