plat/kvm/x86: Map RESERVED regions
When paging is enabled, all memory beyond the kernel image and initrd
is unmapped so that the virtual address space is available again and the
referenced physical memory is managed by the frame allocator. However,
this also removes access to memory-mapped ACPI and APIC regions.
This commit reestablishes these mappings by mapping all reserved
regions provided by multiboot. Unfortunately, multiboot information
is not detailed enough to know what exact type of memory region
we are looking at or what protections are appropriate. For now, we
map these regions as read-only (least privilege).

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>