plat/common: Remove essentials from lcpu assembly
Currently, we include uk/essentials also for assembly files so we get
access to the ALIGN_UP macro. However, this leads to problems when
newlib is included where dependencies cannot be located and we are
in danger of including definitions incompatible with assembly code.
We thus remove the include in this commit and explicitly define ALIGN_UP
for assembly files.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #502