arch/aarch64: Disable outline atomic
Quick-fix based on https://github.com/unikraft/unikraft/issues/319

aarch64-none-elf has the TLS disabled in the gcc configuration,
so aarch64-none-linux-gnu needs to be used for TLS support.

In case it is decided to implement a static version of emutls,
the patch can be reverted to enable the use of the bare-metal toolchain.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>