include/plat: TLS execution helper wrapper
Provides a wrapper that allows a particular function to be executed with
a different TLS pointer enabled. This wrapper solves the problem that
the GCC compiler is not aware when the TLS register is changed. Because
of this, GCC would make incorrect assumptions about TLS variables and
might try to optimize loading and saving a TLS variable at an earlier
or later time within a function, ignoring that the TLS base register
might have changed.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Robert Kuban <robert.kuban@opensynergy.com>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #535