lib/posix-socket: Align posix_socket_driver to 8 bytes in binary
Instances of the `posix_socket_driver` structure (one for each socket
family) are expected to be consecutive in the memory layout. By default,
they are not.

This commit fixes it by aligning the structure to 8 bytes. This ensures
structure instances are consecutively placed in the resulting binary
(and then when loaded in memory).

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>