plat/common/x86: Fix stack alignment
The compiler expects the stack to be always 16-byte aligned
and ending on `...8` when entering a function due to the call
having pushed the return address. So the `pushq` in the
prolog of the callee restores 16-byte alignment. However,
at the moment we do not enforce any particular alignment
for the stack. This can be a problem when the compiler
uses instructions with alignment restrictions like `movaps`.

This commit enforces the required alignment.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
Reviewed-by: Eduard Vintilă <eduard.vintila47@gmail.com>
Approved-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #533