build: Remove optional parameter from fetch rules
This commit removes the dir_with_patches optional parameter from
all the fetch rules that implemented it. The problem of
using patch in the fetch rule is that during fetch, the name of the
base folder in the archive is unknown. This name is needed by the
patch rule in order to know which is the target folder that has
the files that need to be patched. If you want to apply patches to the
folder, please use the patch command after fetching/cloning the folders
that are needed instead.

Signed-off-by: Florin Postolache <florin.postolache80@gmail.com>