Fix types for stat struct
When running the same program using the stat
structure in Linux and with Unikraft the sizes
of the structure's fields were different. This was
caused by the fact that in Linux dev_t was declared
as unsigned long, while in Unikraft it was declared
as __u64. Similarly for ino_t, nlink_t, blkcnt_t.

Signed-off-by: Delia-Maria Pavel <delia_maria.pavel@stud.acs.upb.ro>