support/scripts: qemu-guest: Hugepages for x86 TCG guests
`qemu-guest` uses the qemu64 CPU model for x86 when running the guest
with TCG. This model does not support hugepages in its default setting.
This commit is enabling it by setting the `+pdpe1gb` CPU feature flag.
It is common that this feature is available on recent x86 hardware.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Alexander Jung <alex@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #616