plat/drivers: Fix GICv3 redistributor access
The implementation of the GICv3 driver assumes that the GICR_ registers
access each core's local redistributor. This is not true, as the
redistributor registers of all are mapped in 64KiB adjascent regions,
with GICR_BASE being the offset to the register of the first core. This
results into a bug where all cores update the registributor registers of
the first core.

Update the GIC_RDIST_REG macro to access the right region for a given core,
using the current lcpu index.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #618