build: Add support for long build arguments
Linking of libraries with huge amounts of object files can lead to problems
if paired with a long path to the application directory:

`make[2]: execvp: /bin/bash: Argument list too long`

This commit adds a macro that writes potentially long build arguments into
a file in order to avoid running into the size limits.  The built tool can
then read the arguments from the file, which is an option for most tools in
the gcc toolchain.

This commit only adds this functionality to the `build_cmd` macro.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>