lib/uksched: `uk_sched_start()` returns
This commit changes `uk_sched_start()` so that the resources
of the current context (stack, TLS) continue to execute under
the started scheduler as `init` thread.
Because the thread resources (stack, TLS) are managed external
to the started scheduler, this gives the `init` thread the
opportunity to cleanly free up its resources on exit and
may even enable to start scheduling as part of the `inittab`.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>