lib/posix-socket: Use uk_syscall_e versions for recvfrom / sendto
This is not strictly required. But it is for consistency with other
functions in internal libraries, that also use uk_syscall_r or
uk_syscall_e variants. Those variants are useful to ensure calls to
local implementation. The libc-variants (without the uk_syscall_r /
uk_syscall_e) may clash with libc symbols.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>