scripts/checkpatch: Helper script for checkpatch
Because `checkpatch.pl` does only load a configuration from the current
working directory, this commit adds `support/scripts/checkpatch.uk`. It
is a helper script that executes `checkpatch.pl` with the Unikraft
configuration and can be called from any working directory. This also
supports the use case where the script is symbolically linked to one of
the $PATH directories.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #651