lib/syscall_shim: Use build system's AWK rules for generating sources
Uses the build system's ability for generating source files from AWK
programs. This simplifies the library's `Makefile.uk` and the re-
generation is properly triggered on dependency changes.
Because generated header files are stored in a different destination
path, the same change cannot be performed for these files, at the
moment.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #565