lib/syscall_shim: Use build system's AWK rules for generating sources
Uses the build system's ability for generating source files from AWK
programs. This simplifies the library's `Makefile.uk` and the re-
generation is properly triggered on dependency changes.
Because generated header files are stored in a different destination
path, the same change cannot be performed for these files, at the
moment.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>