lib/nolibc: Use sys/socket.h from Musl
LWIP has been updated to work with Musl. Headers and source code files
from LWIP have been removed since they are part of Musl.

In order to keep nolibc working with LWIP, previous items part of LWIP
now have to be part of nolibc.

This commit removes the current version of sys/socket.h and replaces it
with a fuller version from Musl.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>