lib/nolibc: Add IOV_MAX macro (required by LWIP)
LWIP has been updated to work with Musl. Headers and source code files
from LWIP have been removed since they are part of Musl.

In order to keep nolibc working with LWIP, previous items part of LWIP
now have to be part of nolibc.

This commit adds the IOV_MAX macro in nolibc. The macro was previously
part of LWIP but has now been removed to make it compatible with Musl.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>