lib/nolibc: Add ntohs and htons
LWIP has been updated to work with Musl. Headers and source code
files from LWIP have been removed since they are part of Musl.

In order to keep nolibc working with LWIP, previous items part of
LWIP now have to be part of nolibc.

This commit adds the implementation of `htons()` and `ntohs()`,
functions that had been previously defined in LWIP, but have now been
removed to make it compatible with Musl.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Marco Schlumpp <marco@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #644