lib/vfscore: Initialize stdout, stdin, stderr as early during init
Instead of initializing the default set of file descriptors with a
constructor function, we initialize them during early init. At this
point of boot time, the heap, scheduling, and TLS are initialized.
Because `vfscore` uses the `dup` system call, we introduce this commit
in preparation for the fact that a TLS may be required to perform a
system call.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #565