lib/posix-process: Clone flag `init` and `term` hooks
Similar to the `ctortab` or `inittab`, this commit introduces an inittab
that is used to register handlers that can get involved during a clone
system call depending on given clone flags. This is done for achieving a
proper feature separation with our libraries. For instance,
`CLONE_FILES` specifies if a child shares the file descriptor table
with the parent. Such functionality should be implemented in `vfscore`
(or `posix-fdtab` in the future) instead of `posix-process`.

Checkpatch-Ignore: AVOID_EXTERNS
Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #566