lib/vfscore: Fix stdio.c build warning
When `lib/vfscore` is part of the build, but `lib/syscall_shim` isn't,
there is a warning of implicit declaration of `uk_syscall_r_dup2` in
`lib/vfscore/stdio.c`.

This commit takes care of the warning by declaring `uk_syscall_r_dup2`
when the syscall shim is not part of the build.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #667