build: Always write build command to file
Writes the complete command of build_cmd into a file and executes it.
This avoids the size limit for a single argument (MAX_ARG_STRLEN) relevant
for the execution of complex shell commands by make.
The size of the command should now only be limited by ARG_MAX.
A similar bottleneck is removed from the fixdep invocation.

To ensure build_cmd can be called in immediate and deferred expansion,
a mechanism was added to write to a file only during deferred expansion.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>