lib/posix-futex: Register `set_tid_address` only with `posix-futex`
The `posix-futex` library always registered the `set_tid_address` system
call to `syscall_shim` if `clone` system call support was selected within
`posix-process`. This was independent if the `posix-futex` library was
selected or not. In case, `posix-futex` was not part of the build, the
actual system call symbol was never compiled and caused a linking error.
This commit fixes this by registering the system call only if both
conditions are set.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #681