arch/x86_64: Align the stack to 16 bytes
The System V x86_64 ABI mandates that the stack must be aligned to 16
bytes. Recent compilers have much more aggressive auto-vectorizers and
will also use SIMD instructions to set multiple values on the stack with
one instruction. If the stack is not aligned to 16 bytes, the processor
will generate a general protection fault. This changes the alignment to
the correct 16 bytes.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>