include/uk: Introduce context for init function
This commit introduces a context argument (`struct uk_init_ctx *`) to
the initialization functions. The motivation is that 1) the list of
arguments can be easily extended without changing the init function
signature and 2) we introduce a way to handover boot information
across initalization functions.
This commit adds the argument count (`argc`) and a reference to the
argument vector (`argv`) to `struct uk_init_ctx`. `lib/ukboot` is adopted
accordingly.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1148