include/uk: Extend inittab with termination functions
This commit extends the inittab entries with another function pointer:
a corresponding termination function. The functions are handled as pair
because the idea is that a shutdown means iterating over the table in
reverse order and calling the termination functions. `lib/ukboot` is
adopted accordingly.
Similar to the init functions, a context argument for the termination
functions is introduced, too: `struct uk_term_ctx *`. In its initial
version, it only contains the shutdown request (e.g., `UKPLAT_HALT`,
`UKPLAT_RESTART`).

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1148