lib/uksched: Fix race condition during context switch
The sched refactor effort introduced a race condition: if the current
thread blocks and yields, and an interrupt is triggered right before the
context switch, the interrupt handler tries to wake up the current
thread. The thread will remain blocked (potentially forever) until the
next interrupt fires, because the current thread cannot be added to the
run_queue list. The reason is that the condition
`uk_thread_current() != t` is way too restrictive and does not cover the
corner case where the current thread is just about to be evacuated.

This commit sets a flag on the current thread as soon as the next
thread is scheduled. The interrupt will queue the current thread
if it's also marked as queueable, in addition to it being runnable.

Related-To: 606f5f5
GitHub-Fixes: #643
Checkpatch-Ignore: SINGLE_STATEMENT_DO_WHILE_MACRO
Signed-off-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>