[WILL BE REMOVE] plat/drivers/virtio: Update `virtio_find_vqs()` to find only a particular queue
Previously the `virtio_find_vqs()` used to find the virt queues between
0 and a user provided argument. However this interface assumes that all
required virtqueues i.e. userconfigured + control are sequential.But in
virtio net device the control virt queue is always 2*N and user
configured virtqueues can end at value < 2*N-1. This requies the
separate configuration of control queue.
To incoporate this the `virtio_find_vqs()` function only configures the
queue with provided id, the valid queue id is from 0 to 2*N (control
queue) inclusive.
Along with this the interrupt handler registration is also shifted to
`virtio_pci_add_dev()`. This avoids the multiple calls of this
function because of change in the interface of `virtio_find_vqs()`.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>