build: Add DRUNTIME kconfig option
When building D sources, if no druntime library is supported then we
need to add extra flags to the D compiler (i.e. `-betterC` for DMD or
`-fno-druntime` for gdc).

This commit adds a DRUNTIME hidden kconfig option that will be enabled
by external druntime libraries. If no druntime library is selected, the
appropriate flags will be set.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>