include/uk/plat: Extend memory region descriptor
The commit extends the memory region descriptor in multiple ways to
make it ready to be used within the boot protocol to generically
express the memory layout of the system.

The new structure adds separate virtual and physical address fields
so that scenarios are supported where we don't have a 1:1 mapping
(e.g., when loading the kernel to a different virtual address). Types
are changed accordingly. We also embed the name so that the
descriptor becomes self-contained. The type and flags have been
separated for ease of use and because search semantics for these
are different.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>