include/uk/plat: Extend memory region iteration
The previous search function only accepted exact matches. The new
search function extends this by allowing to specify multiple region
types that are accepted and also select the region flags that should
be incorporated in the search. For example, this allows to list all
readable kernel segments irrespective if they are also writeable or
executable.

For improved performance, the new functions only return a pointer
to the memory region and do not create a copy. While this permits
changing the memory region descriptor, this can be useful in early
boot phases to adapt the region list.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>