include/uk/plat/common: Add memregion list
This commit adds a platform internal memory region list type based
on the public memory region descriptor. The list is effectively a
pre-allocated array that can be filled up to its capacity.
Non-overlapping regions in the list are sorted in ascending order
according to the physical address of the memory regions.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Marco Schlumpp <marco@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Approved-by: Marco Schlumpp <marco@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #722