include/uk/plat/common: Add memregion list
This commit adds a platform internal memory region list type based
on the public memory region descriptor. The list is effectively a
pre-allocated array that can be filled up to its capacity.
Non-overlapping regions in the list are sorted in ascending order
according to the physical address of the memory regions.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>