plat/kvm: Combine multiboot headers
Currently the multiboot definitions span across two headers. The one
that is used in C code only, and one that is safe to include from
assembly. Since we have the __ASSEMBLY__ macro, which is used
extensively in the project, we reduce the number of headers by
combining both multiboot headers into one.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>