plat/common: Introduce boot information
Currently, every boot protocol (e.g., multiboot, Linux Boot Protocol)
must come with their own complete platform initialization as the
basic platform code directly works on the specific boot data
structures. The generalized boot information abstracts away from
the specifics of a certain boot protocol, thereby making it possible to
write generic platform initialization code. Every boot protocol is
thus expected to transform protocol-specific boot information to
the generic form before entering platform initialization code. In case
of chainloading via a custom unikraft loader, the loader may directly
supply boot information in the generalized format.

An important piece of boot information is the memory layout
at the time of boot. That includes the load address of the kernel,
the address of a command line, the load addresses of additional
data blobs (e.g., initrd, devicetree), and where free memory is
located. The boot information structure expresses this as a list
of memory region descriptors. For sanity checks and
compatibility a magic value and version information is provided
at the beginning of the structure.

The boot information is stored in a dedicated section (.uk_bootinfo)
that is initialized with information about the kernel image segments
and their desired access permissions at link-time. To this end, the
linker script needs to call the mkbootinfo.py script and replace the
dummy information in the allocated section with the generated
data.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>