plat/kvm: Back memregions by boot information
Previously, the entries of the memory region list were dynamically
created at runtime based on the section symbols. This commit
replaces the entire implementation and uses the boot information
as source for the list instead. However, this is only done on
x86 for now. We keep the old method for ARM until we have switched
to boot information.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Marco Schlumpp <marco@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Approved-by: Marco Schlumpp <marco@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #722