plat/kvm: Remove omagic flag and remove MULTIBOOT_AOUT_KLUDGE
The omagic flag has caused various issues such as having multiple
segments with different permissions on the same page or producing broken
ELF files. It was necessary because the multiboot header specified the
binary layout in a.out terms. This did not allow any padding between
segments.

With this patch we produce a 32-bit ELF file that is loadable by
QEMU/GRUB without having to resort to MULTIBOOT_AOUT_KLUDGE. This is
necessary because the multiboot protocol expects to a binary that will
start in protected mode. After Unikraft switched to long mode the ELF
file type is not relevant anymore. The `*.dbg` is still a 64-bit ELF
file that can be loaded into GDB to debug a running Unikraft instance in
QEMU. Because multiboot only allows 32-bit ELF files, this also means
that the `*.dbg` cannot be loaded anymore.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Reviewed-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #698