lib/ukboot: Add heap initialization
Previously, the heap has been initialized in the platform code and
information about the heap areas has been communicated via
ALLOCATABLE memory regions. With the new boot information
heap initialization is fully moved to libukboot.

If paging is not enabled, we use the free physical memory areas to
initialize the heap. The platform takes care of only exposing mapped
physical memory. As before, non-contiguous areas are added to the
allocator one-by-one.

If paging is enabled, we create a contiguous mapping at a customizable
base address (default @16GiB) and initialize the allocator with this single
area. As before, using paging allows to access more memory than
mapped by the boot page table.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>