lib/uklock: Provide irqf wrappers for uk_spinlock
Holding a spinlock which is required by an interrupt handler, causes a
deadlock. To avoid this, the programmer must disable the interrupt
before acquiring such spinlocks. This involves saving the interrupt
flags before acquiring the spinlock and restoring them after
releasing the spinlock.

This commit provides versions of the ukplat_spin_* macros for the
uk_spinlock.

Co-authored-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>