[WILL BE REMOVED] lib/uksched/thread.h: Add function to get current thread's stack bottom
To identify the owner of the lock, we need a parameter which is unique
per thread. Since each thread is allocated a unique stack, we can use
the pointer to the stack bottom as owner of the lock.
Also stack is aligned to the 4K boundary, this allows us to use the
bottom 12 bit of the stack pointer to store lock specific info such as
flags

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>