[WILL BE REMOVED] lib/uklock/rwlock: Add SMP safe reader-writer lock
The reader-writer can be acquired by multiple readers simultaneously,
whereas only a single writer can acquire this lock at a given time.

The lock works in two modes: reader and writer.
In each mode, the lower 5 bits stores the flag value, whereas the remaining
bits store the mode-specific info.
In reader mode, the upper bits store the count of readers currently
holding this lock.
In writer mode, the upper bits store the owner of the lock. We use the
stack bottom as the owner since it is unique per thread.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>