[WILL BE REMOVED] lib/uksched/wait: Make wait queues SMP safe
Wait queues uses the linked list to store the threads waiting on a
particular condition. But operations such as add and remove from the
list are not SMP safe. Hence this commit associates a spin lock with
each wait queue. The function which wants to add or remove threads
from wait queue has to hold this lock. This results in mutual exclusion
on SMP system.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>