[WILL BE REMOVED] lib/uklock/rwlock: Decrease number of reader-writer flags from five to four
The reader writer lock used to store its flags in the lower five bits
owner field, This was possible because stack was align to 4K and we had
tweleve free bits in stack pointer. But with new changes in scheduler the
stack is aligned to 16 bytes, hence we only have lower four bits available
for flags.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>